Create AssetRegister components and split data from audit history page
